funciono el codigo del modelo Localproductatribute para claves compuestas

+++ database/migrations/2023_04_05_220727_create_initialinventory_productatribute_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('initialinventory_productatribute', function (Blueprint $table) {
             $table->id();
             //$table->primary('initialinventory_id','productatribute_id');
-            $table->decimal('stock');
+            $table->double('stock');
 
             $table->unsignedBigInteger('initialinventory_id');
             $table->foreign('initialinventory_id')->references('id')->on('initialinventories')->onDelete('cascade');
